move database interfaces and tests to a better place (#1474)
