+++ storagenode/peer.go
@@ -23,6 +23,7 @@ import (
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/pkg/transport"
 	"storj.io/storj/storage"
+	"storj.io/storj/storagenode/bandwidth"
 	"storj.io/storj/storagenode/orders"
 	"storj.io/storj/storagenode/pieces"
 	"storj.io/storj/storagenode/piecestore"
@@ -38,6 +39,12 @@ type DB interface {
 
 	Storage() psserver.Storage
 	Pieces() storage.Blobs
+
+	Orders() orders.DB
+	PieceInfo() pieces.DB
+	CertDB() trust.CertDB
+	Bandwidth() bandwidth.DB
+	UsedSerials() piecestore.UsedSerials
 
 	// TODO: use better interfaces
 	PSDB() *psdb.DB
@@ -96,7 +103,6 @@ type Peer struct {
 
 		Store     *pieces.Store
 		PieceMeta piecestore.PieceMeta
-		Orders    orders.Table
 
 		Endpoint *piecestore.Endpoint
 	}
@@ -204,7 +210,6 @@ func New(log *zap.Logger, full *identity.FullIdentity, db DB, config Config) (*P
 
 		peer.Storage2.Store = pieces.NewStore(peer.Log.Named("pieces"), peer.DB.Pieces())
 		// PieceMeta piecestore.PieceMeta
-		// Orders    orders.Table
 
 		peer.Storage2.Endpoint, err = piecestore.NewEndpoint(
 			peer.Log.Named("piecestore"),
@@ -212,7 +217,7 @@ func New(log *zap.Logger, full *identity.FullIdentity, db DB, config Config) (*P
 			peer.Storage2.Trust,
 			peer.Storage2.Store,
 			peer.Storage2.PieceMeta,
-			peer.Storage2.Orders,
+			peer.DB.Orders(),
 			config.Storage2,
 		)
 		if err != nil {
