+++ storagenode/piecestore/serials.go
@@ -4,6 +4,7 @@
 package piecestore
 
 import (
+	"context"
 	"sync"
 	"time"
 
@@ -12,6 +13,22 @@ import (
 
 	"storj.io/storj/pkg/storj"
 )
+
+// SerialNumberFn is callback from IterateAll
+type SerialNumberFn func(satelliteID storj.NodeID, serialNumber []byte, expiration time.Time)
+
+// UsedSerials is a persistent store for serial numbers.
+// TODO: maybe this should be in orders.UsedSerials
+type UsedSerials interface {
+	// Add adds a serial to the database.
+	Add(ctx context.Context, satelliteID storj.NodeID, serialNumber []byte, expiration time.Time) error
+	// DeleteExpired deletes expired serial numbers
+	DeleteExpired(ctx context.Context, now time.Time) error
+
+	// IterateAll iterates all serials.
+	// Note, this will lock the database and should only be used during startup.
+	IterateAll(ctx context.Context, fn SerialNumberFn) error
+}
 
 type SerialNumbers struct {
 	mu   sync.Mutex
